requirements.txt
python_version < 3.12
numpy==1.26.4
evidently==0.4.5
pandas
fastapi
uvicorn
lightgbm
xgboost
redis
kafka-python
prometheus-client
joblib
requests
shap
matplotlib
